XPipeline - Split LightUnit classes.
