+++ com.unity.render-pipelines.high-definition/Editor/Lighting/LightUnit/HDPunctualLightUnitSlider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 76975b3cce013684cb3d854af3d22d46
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
