+++ com.unity.render-pipelines.high-definition/Editor/Lighting/LightUnit/TemperatureSlider.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 33d6ef6a49896aa44a810a84e3cd394f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
